Move ssh key functions to pkg/pki
Preparation for reuse by the roundtrip mock tests.
