+++ pkg/pki/BUILD.bazel
@@ -6,10 +6,15 @@ go_library(
         "certificate.go",
         "csr.go",
         "privatekey.go",
+        "sshkey.go",
     ],
     importpath = "k8s.io/kops/pkg/pki",
     visibility = ["//visibility:public"],
-    deps = ["//vendor/github.com/golang/glog:go_default_library"],
+    deps = [
+        "//upup/pkg/fi/utils:go_default_library",
+        "//vendor/github.com/golang/glog:go_default_library",
+        "//vendor/golang.org/x/crypto/ssh:go_default_library",
+    ],
 )
 
 go_test(
@@ -17,6 +22,7 @@ go_test(
     srcs = [
         "certificate_test.go",
         "privatekey_test.go",
+        "sshkey_test.go",
     ],
     embed = [":go_default_library"],
 )
